requirements.txt
asgiref==3.8.1
attrs==24.2.0
cattrs==24.1.2
certifi==2024.8.30
charset-normalizer==3.3.2
dj-database-url==2.2.0
Django==5.1.1
django-cors-headers==4.4.0
django-debug-toolbar==4.4.6
django-extensions==3.2.3
django-filter==24.3
django-login-required-middleware==0.9.0
django-phonenumber-field==8.0.0
django-schema-graph==3.1.0
django-ui-forms==1.0
djangorestframework==3.15.2
Faker==30.0.0
idna==3.10
Markdown==3.7
phonenumberslite==8.13.46
pillow==10.4.0
psycopg==3.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3
